# swarm/cmd: Fix some resource leaks in tests