This PR fixes some process/memory and file descriptor leaks. It also disables `TestClientUploadDownloadRawEncrypted` as that test is known to be flaky with `-race`.