@frncmx 2019-03-26
@nonsense 

Let's check if I did not lie. I run: `time go test -run 'TestACT|TestSwarmUp' ./cmd/swarm/`

The left side is `master` and the right side is my branch.

![image](https://user-images.githubusercontent.com/1735667/54988944-c1c47d00-4fb7-11e9-8c32-31ec3be16e74.png)


=> Test run takes the ~same time with my changes but does not leak node(s).