@nonsense 2019-03-29
@frncmx please show a comparison on Travis, not just on your machine. If we introduce slow tests on Travis again, this affects the development time of at least 10 people among the Swarm and Geth teams.