@frncmx 2019-04-04
@nonsense @justelad we are on the same page with:

* I don't want to slow down development/CI run.
* I know about the previous commit improving the test runtime, that's why I referenced it in my commit message. => I spent the time to understand the *context*.
* I understand that we only create random data and the CI runs in a throwaway - 1-time use - VM.
* Considering the context, I fixed the process leak and made *just enough* change to avoid performance degradation. (I think.)

*My mistake*: I did not measure runtime on Travis. Good point! I corrected my mistake. Please, see my results.

---

## Travis measurements

old code: 
```
$ time go test -count 1 -run 'TestACT|TestSwarmUp' ./cmd/swarm/
ok  	github.com/ethereum/go-ethereum/cmd/swarm	46.110s
real	1m57.092s
user	2m31.668s
sys	0m8.184s
```

new code: 
```
$ time go test -count 1 -run 'TestACT|TestSwarmUp' ./cmd/swarm/
ok  	github.com/ethereum/go-ethereum/cmd/swarm	52.019s
real	1m59.327s
user	2m43.108s
sys	0m7.408s
```

**The new test run is 13% slower and adds 7 second to the overall run.**
---
## The problem

*On Travis*: the two tests under discussion run and after they complete they continue to hold 10% of all the available memory until other tests run. (might be ok; I still don't really like 1 test using 10% of all memory available during the whole CI run)

**Locally**: <- this is where my real problem lies
1) I run `go test -run TestACT`; that compiles an ephemeral binary.
2) That binary is run and starts 3 nodes in separate processes. 
3) **The binary exits but those 3 (sometimes just 1) nodes continue running.**
=> I lost 1-3 GB of my computer's memory until reboot or manually killing process. (!)
(If I run tests with `-count 100` at some point my whole computer freezes and I run out of memory.)
---

## The tradeoff 

Note: 
* I made just 1 measurement. I might be wrong, but I don't think I need to run the tests many times to achieve significance.
* However, if I do have to run them many time the new changes will win as the old code will simply run out of memory after 7 to 9 runs.

Leaving as it is:

* Every time I run `TestACT` or `TestSwarmUp` **locally** I'll have some leftover processes eating my memory. I lose 1 GB of memory after each run until I dont clean up. Running tests a 100 or 1000 times is not possilbe if I want to look for flakiness.
* We have enough memory (8 GB) on Travis, so we can afford to lose 1 GB. (That's seem to be false as with `-race` some tests we were not able to run because of memory limitations.)

Accepting the change:
* We accept adding 7s to tests which already run for 46s.
* We free up 1 GB/ 10% memory up on CI.
* Locally I don't have to kill process manually (if I notice even) to keep my memory usage normal.

---

## Suggestion

I'm kinda OK - if that's really the best way - with considering the CI ephemeral and letting *separate* processes to leak. However, my laptop is not ephemeral and we should also consider the local developer experience. I made a mistake not running my tests on Travis, but I corrected that mistake and there seems to be no significant performance degradation on Travis (to my taste). 

I propose that we accept the changes.