@nonsense 2019-04-04
@frncmx nicely done. We still re-use the cluster among the table driven tests, so I guess this is one of the reasons we have only a few seconds slow-down. 7sec. is not at all a lot, so great work.