@frncmx 2019-04-10
submitted upstream: https://github.com/ethereum/go-ethereum/pull/19443

ticket opened for failing race detector job: https://github.com/ethersphere/go-ethereum/issues/1331

@justelad sorry about the delay